demo/requirement.txt
argon2-cffi==19.2.0
asgiref==3.2.3
bcrypt==3.1.7
beautifulsoup4==4.8.1
bs4==0.0.1
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cycler==0.10.0
Django==3.0
django-debug-toolbar==2.1
idna==2.8
lxml==4.4.1
MouseInfo==0.1.2
numpy==1.17.3
pandas==0.25.2
PyAutoGUI==0.9.48
pycparser==2.19
PyGetWindow==0.0.8
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.25
python-dateutil==2.8.0
python-decouple==3.3
PyTweening==1.0.3
pytz==2019.3
requests==2.22.0
six==1.12.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.7
xlrd==1.2.0
